fix(cli): add applies the 0.4.0 transition to templates on 0.4.0+ frameworks
